refactor: use NavItem type from NavItems component
