+++ polaris.shopify.com/src/data/navItems.ts
@@ -13,7 +13,6 @@ interface NavItemWithIcon extends NavItem {
 export const foundationsNavItems: NavItemWithIcon[] = [
   {
     title: "Foundations",
-    url: "/foundations",
     children: [
       {
         title: "Experience values",
